refactor(levelup): merge PitchRestControllerTest
